+++ Demo/src/main/java/com/braintreepayments/demo/PendingRequestStore.java
@@ -5,7 +5,7 @@ import android.content.SharedPreferences;
 
 import com.braintreepayments.api.LocalPaymentPendingRequest;
 import com.braintreepayments.api.PayPalPendingRequest;
-import com.braintreepayments.api.SEPADirectDebitPendingRequest;
+import com.braintreepayments.api.sepadirectdebit.SEPADirectDebitPendingRequest;
 import com.braintreepayments.api.VenmoPendingRequest;
 
 import org.json.JSONException;